recipes/calendar/requirements.txt
Pillow
icalevents
pytz
